# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kapoor Ka Ghulam - Full streaming app API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ProviderSource {
  id: number;
  name: string;
  url: string;
  isDefault: boolean;
  addedAt: string;
}

export interface AddSourceBody {
  /** GitHub author, repo URL, or raw URL */
  input: string;
}

export interface ManifestEntry {
  value: string;
  displayName: string;
  type: string;
  version: string;
  /** @nullable */
  icon?: string | null;
  disabled: boolean;
}

export interface Extension {
  id: number;
  sourceAuthor: string;
  value: string;
  displayName: string;
  type: string;
  /** @nullable */
  icon?: string | null;
  version: string;
  enabled: boolean;
  installedAt: string;
}

export interface InstallExtensionBody {
  sourceId: number;
  value: string;
  displayName: string;
  type: string;
  icon?: string;
  version: string;
}

export interface CatalogItem {
  title: string;
  filter: string;
}

export interface CatalogResponse {
  catalog: CatalogItem[];
  genres: CatalogItem[];
}

export interface Post {
  title: string;
  link: string;
  image: string;
  /** @nullable */
  provider?: string | null;
}

export interface DirectLink {
  title: string;
  link: string;
  /** @nullable */
  type?: string | null;
}

export interface LinkItem {
  title: string;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  episodesLink?: string | null;
  /** @nullable */
  directLinks?: DirectLink[] | null;
}

export interface ContentMeta {
  title: string;
  image: string;
  synopsis: string;
  /** @nullable */
  imdbId?: string | null;
  type: string;
  /** @nullable */
  tags?: string[] | null;
  /** @nullable */
  cast?: string[] | null;
  /** @nullable */
  rating?: string | null;
  linkList: LinkItem[];
}

/**
 * @nullable
 */
export type StreamHeaders = { [key: string]: unknown } | null;

export interface Stream {
  server: string;
  link: string;
  type: string;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  headers?: StreamHeaders;
}

export interface EpisodeLink {
  title: string;
  link: string;
}

export interface SearchResult {
  imdbId: string;
  title: string;
  year: string;
  type: string;
  poster: string;
}

export interface SearchResults {
  results: SearchResult[];
  totalResults: number;
}

export interface ContentInfo {
  imdbId: string;
  title: string;
  year: string;
  type: string;
  poster: string;
  plot: string;
  rating: string;
  genre: string;
  /** @nullable */
  director?: string | null;
  actors: string;
  runtime: string;
  /** @nullable */
  totalSeasons?: string | null;
  language: string;
  country: string;
}

export interface WatchlistItem {
  id: number;
  title: string;
  poster: string;
  link: string;
  provider: string;
  type: string;
  /** @nullable */
  imdbId?: string | null;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  year?: string | null;
  addedAt: string;
}

export interface WatchlistInput {
  title: string;
  poster: string;
  link: string;
  provider: string;
  type: string;
  imdbId?: string;
  rating?: string;
  year?: string;
}

export interface HistoryItem {
  id: number;
  title: string;
  poster: string;
  link: string;
  provider: string;
  type: string;
  /** @nullable */
  imdbId?: string | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  episodeTitle?: string | null;
  watchedAt: string;
}

export interface HistoryInput {
  title: string;
  poster: string;
  link: string;
  provider: string;
  type: string;
  imdbId?: string;
  progress?: number;
  duration?: number;
  episodeTitle?: string;
}

export interface AppSettings {
  /** @nullable */
  activeExtId: number | null;
  theme: string;
}

export interface AppSettingsInput {
  activeExtId?: number;
  theme?: string;
}

export interface UserStats {
  watchlistCount: number;
  historyCount: number;
  extensionsCount: number;
  sourcesCount: number;
}

export type FetchManifestParams = {
sourceId: number;
};

export type GetCatalogParams = {
extId: number;
};

export type GetPostsParams = {
extId: number;
filter: string;
page?: number;
};

export type SearchPostsParams = {
extId: number;
q: string;
page?: number;
};

export type GetMetaParams = {
extId: number;
link: string;
};

export type GetStreamsParams = {
extId: number;
link: string;
type?: string;
};

export type GetEpisodesParams = {
extId: number;
url: string;
};

export type SearchContentParams = {
q: string;
type?: SearchContentType;
};

export type SearchContentType = typeof SearchContentType[keyof typeof SearchContentType];


export const SearchContentType = {
  movie: 'movie',
  series: 'series',
  episode: 'episode',
} as const;

export type GetContentInfoParams = {
imdbId: string;
};