# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Kapoor Ka Ghulam - Full streaming app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sources:
    get:
      operationId: getSources
      tags: [sources]
      summary: Get all provider sources
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSource"

    post:
      operationId: addSource
      tags: [sources]
      summary: Add a provider source (GitHub URL or author name)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSourceBody"
      responses:
        "201":
          description: Source added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderSource"
  /sources/{id}:
    delete:
      operationId: removeSource
      tags: [sources]
      summary: Remove a provider source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /sources/{id}/default:
    put:
      operationId: setDefaultSource
      tags: [sources]
      summary: Set a source as default
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Manifest / Extensions ────────────────────────────────────
  /manifest:
    get:
      operationId: fetchManifest
      tags: [extensions]
      summary: Fetch manifest from a provider source
      parameters:
        - name: sourceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Manifest entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManifestEntry"
...
  /extensions/install:
    post:
      operationId: installExtension
      tags: [extensions]
      summary: Install an extension (downloads and caches JS modules)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallExtensionBody"
      responses:
        "201":
          description: Installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Extension"
  /extensions/{id}:
    delete:
      operationId: uninstallExtension
      tags: [extensions]
      summary: Uninstall an extension
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Uninstalled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Content via Extensions ───────────────────────────────────
  /catalog:
    get:
      operationId: getCatalog
      tags: [content]
      summary: Get catalog categories for an extension
      parameters:
        - name: extId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Catalog categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogResponse"
  /posts:
    get:
      operationId: getPosts
      tags: [content]
      summary: Get posts for a catalog filter
      parameters:
        - name: extId
          in: query
          required: true
          schema:
            type: integer
        - name: filter
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /posts/search:
    get:
      operationId: searchPosts
      tags: [content]
      summary: Search posts in an extension
      parameters:
        - name: extId
          in: query
          required: true
          schema:
            type: integer
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /meta:
    get:
      operationId: getMeta
      tags: [content]
      summary: Get metadata for a content item from extension
      parameters:
        - name: extId
          in: query
          required: true
          schema:
            type: integer
        - name: link
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Content metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentMeta"
  /streams:
    get:
      operationId: getStreams
      tags: [content]
      summary: Get stream URLs for a content item
      parameters:
        - name: extId
          in: query
          required: true
          schema:
            type: integer
        - name: link
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Stream URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stream"
  /episodes:
    get:
      operationId: getEpisodes
      tags: [content]
      summary: Get episodes for a series season
      parameters:
        - name: extId
          in: query
          required: true
          schema:
            type: integer
        - name: url
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Episodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EpisodeLink"

  # ── OMDB Search ───────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── Sources ─────────────────────────────────────────────────
    ProviderSource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        isDefault:
          type: boolean
        addedAt:
          type: string
      required: [id, name, url, isDefault, addedAt]
    AddSourceBody:
      type: object
      properties:
        input:
          type: string
          description: GitHub author, repo URL, or raw URL
      required: [input]

    # ── Extensions ───────────────────────────────────────────────
    ManifestEntry:
      type: object
      properties:
        value:
          type: string
        displayName:
          type: string
        type:
          type: string
        version:
          type: string
        icon:
          type: ["string", "null"]
        disabled:
          type: boolean
      required: [value, displayName, type, version, disabled]
    Extension:
      type: object
      properties:
        id:
          type: integer
        sourceAuthor:
          type: string
        value:
          type: string
        displayName:
          type: string
        type:
          type: string
        icon:
          type: ["string", "null"]
        version:
          type: string
        enabled:
          type: boolean
        installedAt:
          type: string
      required: [id, sourceAuthor, value, displayName, type, version, enabled, installedAt]
    InstallExtensionBody:
      type: object
      properties:
        sourceId:
          type: integer
        value:
          type: string
        displayName:
          type: string
        type:
          type: string
        icon:
          type: string
        version:
          type: string
      required: [sourceId, value, displayName, type, version]

    # ── Content ──────────────────────────────────────────────────
    CatalogItem:
      type: object
      properties:
        title:
          type: string
        filter:
          type: string
      required: [title, filter]
    CatalogResponse:
      type: object
      properties:
        catalog:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItem"
        genres:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItem"
      required: [catalog, genres]
    Post:
      type: object
      properties:
        title:
          type: string
        link:
          type: string
        image:
          type: string
        provider:
          type: ["string", "null"]
      required: [title, link, image]
    DirectLink:
      type: object
      properties:
        title:
          type: string
        link:
          type: string
        type:
          type: ["string", "null"]
      required: [title, link]
    LinkItem:
      type: object
      properties:
        title:
          type: string
        quality:
          type: ["string", "null"]
        episodesLink:
          type: ["string", "null"]
        directLinks:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/DirectLink"
      required: [title]
    ContentMeta:
      type: object
      properties:
        title:
          type: string
        image:
          type: string
        synopsis:
          type: string
        imdbId:
          type: ["string", "null"]
        type:
          type: string
        tags:
          type: ["array", "null"]
          items:
            type: string
        cast:
          type: ["array", "null"]
          items:
            type: string
        rating:
          type: ["string", "null"]
        linkList:
          type: array
          items:
            $ref: "#/components/schemas/LinkItem"
      required: [title, image, synopsis, type, linkList]
    Stream:
      type: object
      properties:
        server:
          type: string
        link:
          type: string
        type:
          type: string
        quality:
          type: ["string", "null"]
        headers:
          type: ["object", "null"]
      required: [server, link, type]
    EpisodeLink:
      type: object
      properties:
        title:
          type: string
        link:
          type: string
      required: [title, link]

    # ── OMDB ─────────────────────────────────────────────────────
...